refactor(api): перенос на ts
