+++ 10-weather-typescript/services/api/rest/api.service.ts
@@ -1,14 +1,23 @@
-import axios from "axios";
-import { getWeatherApiInfo } from "../../log/rest/log.service.js";
+import axios, { AxiosError } from "axios";
+import { getWeatherApiInfo } from "../../log/rest/log.service";
 import {
   getCities,
   getLanguage,
   getToken,
-} from "../../storage/storage.service.js";
-import { DEFAULT_LANGUAGE, WEATHER_URL, getIcon } from "../shared/helpers.js";
-import { HttpUtils } from "../shared/httpUtils.js";
+} from "../../storage/storage.service";
+import { DEFAULT_LANGUAGE, WEATHER_URL, getIcon } from "../shared/helpers";
+import { HttpUtils } from "../shared/httpUtils";
 
-const getWeather = async ({ city, fromQuery: { token, lang } }) => {
+const getWeather = async ({
+  city,
+  fromQuery: { token, lang },
+}: {
+  city: string;
+  fromQuery: {
+    token: string;
+    lang: "ru" | "eng";
+  };
+}) => {
   const OPEN_WEATHER_MAP_API_KEY = token || (await getToken());
   const LANGUAGE = lang || (await getLanguage()) || DEFAULT_LANGUAGE;
   if (!OPEN_WEATHER_MAP_API_KEY) {
@@ -28,6 +37,8 @@ const getWeather = async ({ city, fromQuery: { token, lang } }) => {
 
 const getForecast = async ({
   fromQuery: { token, lang, cities: citiesFromQuery },
+}: {
+  fromQuery: { token: string; lang: "ru" | "eng"; cities: string[] };
 }) => {
   try {
     const cities = citiesFromQuery || (await getCities());
@@ -44,15 +55,20 @@ const getForecast = async ({
         getWeatherApiInfo(weather, getIcon(weather.weather[0].icon))
       )
       .join("\n");
-  } catch (err) {
-    if (err?.response?.status == 404) {
-      throw new Error("Город не найден");
-    } else if (err?.response?.status == 401) {
-      throw new Error("Не авторизован. Установите токен");
-    } else {
+  } catch (err: unknown) {
+    if (err instanceof AxiosError) {
+      if (err?.response?.status == 404) {
+        throw new Error("Город не найден");
+      } else if (err?.response?.status == 401) {
+        throw new Error("Не авторизован. Установите токен");
+      } else {
+        throw new Error(err.message);
+      }
+    } else if (err instanceof Error) {
       throw new Error(err.message);
     }
   }
 };
 
 export { getForecast, getIcon, getWeather };
+
