+++ 10-weather-typescript/services/api/shared/helpers.ts
