+++ 10-weather-typescript/services/api/shared/httpUtils.ts
